backend/requirements.txt
firebase_admin==6.5.0
Flask==3.0.3
gTTS==2.5.1
moviepy==1.0.3
openai==1.26.0
parse==1.20.1
python-dotenv==1.0.1
Requests==2.31.0
soundfile==0.12.1
